apps/api/pyproject.toml
[project]
name = "skillhub-api"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "apiflask>=1.3",
    "flask-cors",
    "gunicorn",
    "pydantic>=2.0",
    "pydantic-settings",
    "sqlalchemy>=2.0",
    "alembic",
    "psycopg2-binary",
    "pyjwt",
    "httpx",
    "redis>=5.0",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-flask",
    "opentelemetry-instrumentation-sqlalchemy",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov"]
